Avoid lazy imports in new tests
